+++ CHANGELOG.md
@@ -103,6 +103,7 @@ It includes many new rules, including all new techniques introduced in MITRE ATT
 - show-features: don't show features from library functions #569 @williballenthin
 - linter: summarize results at the end #571 @williballenthin
 - meta: added `library_functions` field, `feature_counts.functions` does not include library functions any more #562 @mr-tz
+- linter: check for `or` with always true child statement, e.g. `optional`, colors #348 @mr-tz
 
 ### Development